-MapBox Control The MapBox Control nows render the Layers Asynchronsously. There is a small progress bar in the upper-left corner that is running while the Render process is active With this method, when there are "huge" / "heavy" layers (Big shapefiles with lots of rendering options, ex) the render is running in a separate thread and frees the UI thread. In the MapBox control, the DynamicLayers are also considered to make the rendering of the 2 images: StaticBaseImage + DynamicImage. Then there is a merge of the 2 images.
- DynamicLayerCollection
There is only one global timer for all the DynamicLayercollections. otherwise, if there was a timer for each Map, when there was being rendered 10 Maps at the same time, the CPU will be all the time being used. With only one timer, the same timer is used to render all the DynamicLayerCollections in each Map

- ShapeFile Providers/ShapeFile.cs
Inseted a local cache to have the recent obtained FeatureRows. Usually in a Zoom/Pan operation there are some feature that remains in the view. With this cache there is a performance improvement

- AsyncTileLayer
With this new TileLayer, the tiles are rendered in asynchronously, and the image is rendered ala-GoogleMaps instead of the full image at the same time

- winFormSamples
I've created a new Form (MDI) that shows the example of having n-Maps in the same Window and that can be used to show the async problem.